requirements.txt
gradio
streamlit
requests
beautifulsoup4
selenium
pandas
numpy
python-dotenv
pydantic
openai
anthropic
apify-client
plotly
